Campground premission

+++ routes/campgrounds.js
@@ -5,6 +5,7 @@ const ExpressError = require('../utils/ExpressError');
 const Campground = require('../models/campground');
 const { campgroundSchema } = require('../schemas.js');
 const isLoggedIn = require('../middleware');
+const { raw } = require('express');
 
 const validateCampground = (req, res, next) => {
   const { error } = campgroundSchema.validate(req.body);
@@ -61,10 +62,15 @@ router.get(
   '/:id/edit',
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if (!campground) {
       req.flash('error', 'Cannot find that campground!');
       return res.redirect('/campgrounds');
+    }
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You not have premission to do that!');
+      return res.redirect(`/campgrounds/${id}`);
     }
     res.render('campgrounds/edit', { campground });
   })
@@ -76,11 +82,16 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You not have premission to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const updatedCampground = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
     req.flash('success', 'Successfully updated campground!');
-    res.redirect(`/campgrounds/${campground._id}`);
+    res.redirect(`/campgrounds/${updatedCampground._id}`);
   })
 );
 
@@ -89,6 +100,11 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You not have premission to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground!');
     res.redirect(`/campgrounds`);
